Cart

+++ Nhom11/Controllers/CartsController.cs
@@ -29,34 +29,42 @@ namespace Nhom11.Controllers
             var cart=_context.Carts.Include(c=>c.Account).Include(c=>c.Product).Where(c=>c.Account.Username==username);
             var accountid=_context.Accounts.FirstOrDefault(a=>a.Username==username).Id;
             var total=cart.Sum(c=>c.Product.Price*c.Quantity);
-            Invoice invoice = new Invoice()
+            if (cart != null)
             {
-                Code = DateTime.Now.ToString("yyyyMMddhhmmss"),
-                AccountId=accountid,
-                IssuedDate=DateTime.Now,
-                ShippingAddress=ShippingAddress,
-                ShippingPhone=ShippingPhone,
-                Total=total,
-                Status=true,
-            };
-            _context.Invoices.Add(invoice);
-            _context.SaveChanges();
-            foreach(var item in cart)
-            {
-                InvoiceDetail detail = new InvoiceDetail()
+                Invoice invoice = new Invoice()
                 {
-                    InvoiceId = invoice.Id,
-                    ProductId = item.ProductId,
-                    Quantity = item.Quantity,
-                    UnitPrice = item.Product.Price
+                    Code = DateTime.Now.ToString("yyyyMMddhhmmss"),
+                    AccountId = accountid,
+                    IssuedDate = DateTime.Now,
+                    ShippingAddress = ShippingAddress,
+                    ShippingPhone = ShippingPhone,
+                    Total = total,
+                    Status = true,
                 };
-                _context.InvoiceDetails.Add(detail);
-                _context.Carts.Remove(item);
-                item.Product.Stock -= item.Quantity;
-                _context.Products.Update(item.Product);
+                _context.Invoices.Add(invoice);
+                _context.SaveChanges();
+                foreach (var item in cart)
+                {
+                    InvoiceDetail detail = new InvoiceDetail()
+                    {
+                        InvoiceId = invoice.Id,
+                        ProductId = item.ProductId,
+                        Quantity = item.Quantity,
+                        UnitPrice = item.Product.Price
+                    };
+                    _context.InvoiceDetails.Add(detail);
+                    _context.Carts.Remove(item);
+                    item.Product.Stock -= item.Quantity;
+                    _context.Products.Update(item.Product);
+                }
+                _context.SaveChanges();
+                return RedirectToAction("Index", "Products");
             }
-            _context.SaveChanges();
-            return RedirectToAction("Index","Products");
+            else
+            {
+                return RedirectToAction("Index");
+            }
+            
         }
         [HttpPost]
         public async Task<IActionResult> Add(int ProductId,int Quantity,int Size)
